pyproject.toml
[tool.poetry]
name = "pysdrweb"
version = "0.3.2"
description = "Server to host FM Radio via RTL-SDR utilities."
authors = ["Aaron Gibson <eulersidcrisis@yahoo.com>"]
license = "MIT License"
readme = "README.md"
repository = "https://github.com/eulersIDcrisis/pysdrweb"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Telecommunications Industry",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX",
    "Topic :: Communications :: Ham Radio",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Players"
]


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/eulersIDcrisis/pysdrweb/issues"

[tool.poetry.dependencies]
python = ">=3.8,<4"
tornado = "^6.1"
PyYAML = "^6.0"
click = "^8.0.4"
SoundFile = "^0.10.3"
lameenc = "1.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
    sdrfm_server = 'pysdrweb.fmserver.server:fm_server_command'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
